tenants-backend/requirements2.txt
backports.tarfile==1.2.0
bleach==6.2.0
brotli==1.1.0
channels-redis==4.2.1
chardet==5.2.0
colorama==0.4.6
daphne==4.2.1
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==25.1
django-ipware==7.0.1
django-tenant-users==2.1.1
djangorestframework-simplejwt==5.5.0
factory-boy==3.3.3
hashids==1.3.1
hypothesis==6.136.7
importlib-metadata==8.0.0
jaraco.collections==5.1.0
markdown==3.8.2
pandas==2.3.0
pip-chill==1.0.3
platformdirs==4.2.2
psycopg2==2.9.10
pyopenssl==25.1.0
pypdf==6.0.0
pytest-django==4.11.1
reportlab==4.4.2
service-identity==24.2.0
tomli==2.0.1
watchfiles==1.1.0
weasyprint==66.0
zope.interface==7.2
zopfli==0.2.3.post1
